==> data_processing/4-stretch-goals.qmd <==
---
title: "4-stretch-goals"
format: html
knitr:
  opts_chunk:
    fig.path: "../images/"
---

These are some aspects of my work that I ultimately decided to move to a sort of "stretch goals" of the project. Elements I was interested in but weren't necessarily pertinent to the analysis I was focused one. Or, in the case of emojis usage, elements I Was interested in but didn't strictly have the R know-how to get working exactly how I wanted. 

I also used a specific emoji usage measurement that would not render from a qmd with gfm format, and required that I use html format. Rather than do that to my entire analysis notebook, I decided to isolate the emoji usage details here instead. 

```{r}
#| label: setup
#| include: false
library(tidyverse)
library(tidytext)
library(ggplot2)
library(highcharter)
library(tidyr)
library(scales)
```

**Note**: Please run [0-data-exploration](https://github.com/Data-Sci-2025/Steam-Reviews-Analysis/blob/main/data_processing/0-data-exploration.qmd) and [1-data-cleanup](https://github.com/Data-Sci-2025/Steam-Reviews-Analysis/blob/main/data_processing/1-data-cleanup.qmd). I exported this .csv file from the midpoint of the [2-data-analysis](https://github.com/Data-Sci-2025/Steam-Reviews-Analysis/blob/main/data_processing/2-data-analysis.qmd) file, after adding word types.


```{r}
#| label: readingin
reviews_df <- read_csv("../private/reviews_emojis.csv", show_col_types = FALSE)
```


## Reviews over time

```{r}
#| label: failed-plot
#| eval: false
ggplot(reviews_df, aes(x=Date_Published)) +   stat_count(geom='line', aes(y=..count..))
```

I wanted to make a line chart here per game title of the average review count, all overlapping. I couldn't quite get it working and ran out of time. I would like to see how games in the positive Steam rankings (likely) have some longevity in incoming reviews, while more negative games likely just fall off after a short while. 



## Emojis

A lot of work went into keeping those emojis, so I want to take a look at them! 

```{r}
#| label: emoji-str
emoji_pattern <- "[\\\U0001F600-\\\U0001F64F\\\U0001F300-\\\U0001F5FF\\\U0001F900-\\\U0001F9FF\\\U00002600-\\\U000027BF]"
```


```{r}
#| label: extract-emojis
emojis <- reviews_df |>
  filter(grepl(emoji_pattern, types, perl=TRUE)) |>
  mutate(types = str_remove_all(types, "\\w|,|'|‘|’
                                 %|\\.|’|\\\036|－|❝|❞
                                 |‑"))

emojis
```


Only 1,622 of our reviews contain some kind of emoji usage. Not as many as I expected!
Note that it's 528 here because I believe the file was exported on the downsampled data. 

```{r}
#| label: GetTypes
remove_dupes <- function(x) {
  #split tokens to be iterated over
  words <- strsplit(x, " ")[[1]]
  #apply the unique() function per row
  unique_words <- unique(words)
  #re-collapse into rows
  paste(unique_words, collapse = " ")
}
```


(sorry that these don't look beautiful in the .md render, the highcharter graph below won't render unless the format is html)


```{r}
#| label: emojis-only
emoji_df <- emojis |>
  mutate(types = sapply(emojis$types, remove_dupes)) |>
  count(types, sort=TRUE)

emoji_df
```


```{r}
#| label: plot-emoji
max_count <- emoji_df |>
  pull(n) |>
  max()


hchart(
  emoji_df, 
  "bar",
  hcaes(types, n), 
  name = "Count",
  # dataLabels = list(enabled = TRUE, style = list(fontWeight = "normal"))
  ) |>
  hc_xAxis(
    min = 0,
    max = 30,
    scrollbar = list(enabled = TRUE)
    ) |>
  hc_yAxis(
    max = max_count,
    title = list(
      text = "Count",
      align = "high"
      )
    ) |>
  hc_tooltip(
    headerFormat = "{point.key}",
    pointFormat = " {point.y}"
  ) |>
  hc_size(height = 700)
```


Cool [source](https://jkunst.com/blog/posts/2021-01-10-using-emoji-with-highcharterhighcharts/) for the emoji finding & charting code. However there were still too many errors for my taste in how the emojis were splitting up. Thumbs up appears twice for some reason, emojis are appearing in combination with each other.... It would take some more serious treatment (and time) to get this looking how I want it to look. 



## Language Choices

I wanted to use the [sentimentr](https://towardsdatascience.com/how-to-catch-profanity-with-r-4a21e024d4db/) package which comes loaded with multiple profanity lexicons to detect foul language throughout these reviews! How common is swearing in reviews? Is it more common in negative reviews than positive? What kind of swears are people using?
